add helper common functions for reading ingresses and extract them from file
