+++ pkg/i2gw/ingress2gateway.go
@@ -17,18 +17,10 @@ limitations under the License.
 package i2gw
 
 import (
-	"bytes"
 	"context"
-	"errors"
 	"fmt"
-	"io"
-	"os"
 
-	networkingv1 "k8s.io/api/networking/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
-	kubeyaml "k8s.io/apimachinery/pkg/util/yaml"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 )
@@ -98,14 +90,6 @@ func readProviderResourcesFromCluster(ctx context.Context, providerByName map[Pr
 	return nil
 }
 
-func ConstructIngressesFromCluster(ctx context.Context, cl client.Client, ingressList *networkingv1.IngressList) error {
-	err := cl.List(ctx, ingressList)
-	if err != nil {
-		return fmt.Errorf("failed to get ingresses from the cluster: %w", err)
-	}
-	return nil
-}
-
 // constructProviders constructs a map of concrete Provider implementations
 // by their ProviderName.
 func constructProviders(conf *ProviderConf, providers []string) (map[ProviderName]Provider, error) {
@@ -121,84 +105,6 @@ func constructProviders(conf *ProviderConf, providers []string) (map[ProviderNam
 	}
 
 	return providerByName, nil
-}
-
-// ExtractObjectsFromReader extracts all objects from a reader,
-// which is created from YAML or JSON input files.
-// It retrieves all objects, including nested ones if they are contained within a list.
-// The function takes a namespace parameter to optionally return only namespaced resources.
-func ExtractObjectsFromReader(reader io.Reader, namespace string) ([]*unstructured.Unstructured, error) {
-	d := kubeyaml.NewYAMLOrJSONDecoder(reader, 4096)
-	var objs []*unstructured.Unstructured
-	for {
-		u := &unstructured.Unstructured{}
-		if err := d.Decode(&u); err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			return objs, fmt.Errorf("failed to unmarshal manifest: %w", err)
-		}
-		if u == nil {
-			continue
-		}
-		if namespace != "" && u.GetNamespace() != namespace {
-			continue
-		}
-		objs = append(objs, u)
-	}
-
-	finalObjs := []*unstructured.Unstructured{}
-	for _, obj := range objs {
-		tmpObjs := []*unstructured.Unstructured{}
-		if obj.IsList() {
-			err := obj.EachListItem(func(object runtime.Object) error {
-				unstructuredObj, ok := object.(*unstructured.Unstructured)
-				if ok {
-					tmpObjs = append(tmpObjs, unstructuredObj)
-					return nil
-				}
-				return fmt.Errorf("resource list item has unexpected type")
-			})
-			if err != nil {
-				return nil, err
-			}
-		} else {
-			tmpObjs = append(tmpObjs, obj)
-		}
-		finalObjs = append(finalObjs, tmpObjs...)
-	}
-
-	return finalObjs, nil
-}
-
-// ConstructIngressesFromFile reads the inputFile in either json/yaml formats,
-// then deserialize the file into Ingresses resources.
-// All ingresses will be pushed into the supplied IngressList for return.
-func ConstructIngressesFromFile(l *networkingv1.IngressList, inputFile string, namespace string) error {
-	stream, err := os.ReadFile(inputFile)
-	if err != nil {
-		return err
-	}
-
-	reader := bytes.NewReader(stream)
-	objs, err := ExtractObjectsFromReader(reader, namespace)
-	if err != nil {
-		return err
-	}
-
-	for _, f := range objs {
-		if !f.GroupVersionKind().Empty() && f.GroupVersionKind().Kind == "Ingress" {
-			var i networkingv1.Ingress
-			err = runtime.DefaultUnstructuredConverter.
-				FromUnstructured(f.UnstructuredContent(), &i)
-			if err != nil {
-				return err
-			}
-			l.Items = append(l.Items, i)
-		}
-
-	}
-	return nil
 }
 
 func aggregatedErrs(errs field.ErrorList) error {
