+++ pkg/i2gw/providers/common/testdata/input-file.yaml
