+++ pkg/i2gw/providers/ingressnginx/resource_reader.go
@@ -17,16 +17,10 @@ limitations under the License.
 package ingressnginx
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"os"
 
 	"github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw"
 	"github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/common"
-	networkingv1 "k8s.io/api/networking/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 )
 
 // converter implements the i2gw.CustomResourceReader interface.
@@ -44,48 +38,21 @@ func newResourceReader(conf *i2gw.ProviderConf) *resourceReader {
 func (r *resourceReader) readResourcesFromCluster(ctx context.Context) (*storage, error) {
 	storage := newResourcesStorage()
 
-	var ingressList networkingv1.IngressList
-	err := r.conf.Client.List(ctx, &ingressList)
+	ingresses, err := common.ReadIngressesFromCluster(ctx, r.conf.Client, NginxIngressClass)
 	if err != nil {
-		return nil, fmt.Errorf("failed to get ingresses from the cluster: %w", err)
+		return nil, err
 	}
-
-	for i, ingress := range ingressList.Items {
-		if common.GetIngressClass(ingress) != NginxIngressClass {
-			continue
-		}
-		storage.Ingresses[types.NamespacedName{Namespace: ingress.Namespace, Name: ingress.Name}] = &ingressList.Items[i]
-	}
-
+	storage.Ingresses = ingresses
 	return storage, nil
 }
 
-func (r *resourceReader) readResourcesFromFile(_ context.Context, filename string) (*storage, error) {
+func (r *resourceReader) readResourcesFromFile(filename string) (*storage, error) {
 	storage := newResourcesStorage()
-	stream, err := os.ReadFile(filename)
+
+	ingresses, err := common.ReadIngressesFromFile(filename, r.conf.Namespace, NginxIngressClass)
 	if err != nil {
-		return nil, fmt.Errorf("failed to read file %v: %w", filename, err)
+		return nil, err
 	}
-
-	unstructuredObjects, err := i2gw.ExtractObjectsFromReader(bytes.NewReader(stream), r.conf.Namespace)
-	if err != nil {
-		return nil, fmt.Errorf("failed to extract objects: %w", err)
-	}
-
-	for _, f := range unstructuredObjects {
-		if !f.GroupVersionKind().Empty() && f.GroupVersionKind().Kind == "Ingress" {
-			var i networkingv1.Ingress
-			err = runtime.DefaultUnstructuredConverter.
-				FromUnstructured(f.UnstructuredContent(), &i)
-			if err != nil {
-				return nil, err
-			}
-			if common.GetIngressClass(i) != NginxIngressClass {
-				continue
-			}
-			storage.Ingresses[types.NamespacedName{Namespace: i.Namespace, Name: i.Name}] = &i
-		}
-
-	}
+	storage.Ingresses = ingresses
 	return storage, nil
 }
